Create allow-unsafe-urls.php
Workaround to import images WordPress Woocommerce: Error: A valid URL was not provided.

While importing products using import/export tools sometimes you get an error, by adding this to the functions.php of the Theme you can workaround. It is recommended to remove this when you are finished importing products and images.